Part 1: off-chain RWA price feed receiver via Pyth Lazer WebSocket
Stream XAU/USD and XAG/USD from Pyth Lazer in solana binary format,
logging JSON-parsed feed data and the Ed25519-signed payload used
as redeemer in Cardano transactions.
