+++ lazer/cardano/lazer-rwa/package.json
@@ -1,22 +1,18 @@
 {
-  "name": "cardano-pyth-draft",
+  "name": "lazer-rwa",
   "version": "1.0.0",
   "type": "module",
   "scripts": {
-    "fetch-prices": "npx tsx src/fetch_prices.ts",
+    "fetch-prices": "npx tsx src/fetch_prices_lazer_rwa.ts",
     "submit-tx": "npx tsx src/submit_tx.ts"
   },
   "dependencies": {
     "@evolution-sdk/evolution": "^0.3.30",
     "@pythnetwork/pyth-lazer-cardano-js": "^0.1.0",
-    "@pythnetwork/pyth-lazer-sdk": "^5.2.0",
-    "@types/yargs": "^17.0.35",
-    "bip39": "^3.1.0",
-    "yargs": "^18.0.0"
+    "@pythnetwork/pyth-lazer-sdk": "^5.2.0"
   },
   "devDependencies": {
     "@types/node": "^22.10.0",
-    "ts-node": "^10.9.2",
     "tsx": "^4.21.0",
     "typescript": "^5.7.2"
   }
